Fix Agent crash in FlowStatsCollector::DeleteEvent
When a flow is removed from Flow tree of FlowStatsCollector, reset the
FlowStatsCollector pointer of flow as it is possible that FlowStatsCollector
can be deleted. Also FlowStatsCollector now ignores delete request received,
if the flow does not have valid FlowStatsCollector pointer (duplicate delete)

Closes-Bug: #1577084

+++ src/vnsw/agent/pkt/flow_table.cc
@@ -130,9 +130,13 @@ FlowEntry *FlowTable::Find(const FlowKey &key) {
 }
 
 void FlowTable::Copy(FlowEntry *lhs, FlowEntry *rhs, bool update) {
-    RevFlowDepParams params;
-    lhs->RevFlowDepInfo(&params);
-    DeleteFlowInfo(lhs, params);
+    /* If flow is delete-marked we would have already sent a DELETE msg to
+     * FlowMgmt */
+    if (!lhs->deleted()) {
+        RevFlowDepParams params;
+        lhs->RevFlowDepInfo(&params);
+        DeleteFlowInfo(lhs, params);
+    }
     if (rhs)
         lhs->Copy(rhs, update);
 }
